ROBOT-1118: Split up Robot.py slightly
