+++ mecademic/Robot.py
@@ -12,6 +12,8 @@ import contextlib
 import os
 
 from .mx_robot_def import *
+from .Common import *
+from .Logger import *
 
 CHECKPOINT_ID_MAX_PRIVATE = 8191  # Max allowable checkpoint id, inclusive
 
@@ -19,246 +21,6 @@ TERMINATE = '--terminate--'
 
 GRIPPER_OPEN = True
 GRIPPER_CLOSE = False
-
-
-class MecademicException(Exception):
-    """Base exception class for Mecademic-related exceptions.
-
-    """
-    pass
-
-
-class InvalidStateError(MecademicException):
-    """The internal state of the instance is invalid.
-
-    """
-    pass
-
-
-class CommunicationError(MecademicException):
-    """There is a communication issue with the robot.
-
-    """
-    pass
-
-
-class DisconnectError(MecademicException):
-    """A non-nominal disconnection has occurred.
-
-    """
-    pass
-
-
-class InterruptException(MecademicException):
-    """An event has encountered an error. Perhaps it will never be set.
-
-    """
-    pass
-
-
-class InterruptableEvent:
-    """Extend default event class to also be able to unblock and raise an exception.
-
-    Attributes
-    ----------
-    id : int or None
-        Id for event.
-    _event : event object
-        A standard event-type object.
-    _lock : lock object
-        Used to ensure atomic operations.
-    _interrupted : boolean
-        If true, event is in an error state.
-
-    """
-    def __init__(self, id=None):
-        self.id = id
-        self._event = threading.Event()
-        self._lock = threading.Lock()
-        self._interrupted = False
-
-    def wait(self, timeout=None):
-        """Block until event is set or should raise an exception.
-
-        Attributes
-        ----------
-        timeout : float
-            Maximum duration to wait in seconds.
-
-        Return
-        ------
-        success : boolean
-            False if event timed out, true otherwise.
-
-        """
-        success = self._event.wait(timeout=timeout)
-        if self._interrupted:
-            raise InterruptException('Event received exception, possibly because event will never be triggered.')
-        return success
-
-    def set(self):
-        """Set the event and unblock all waits.
-
-        """
-        with self._lock:
-            self._event.set()
-
-    def abort(self):
-        """Unblock any waits and raise an exception.
-
-        """
-        with self._lock:
-            if not self._event.is_set():
-                self._interrupted = True
-                self._event.set()
-
-    def clear(self):
-        """Reset the event to its initial state.
-
-        """
-        with self._lock:
-            self._interrupted = False
-            self._event.clear()
-
-    def is_set(self):
-        """Checks if the event is set.
-
-        Return
-        ------
-        boolean
-            False if event is not set or instance should '_interrupted'. True otherwise.
-
-        """
-        with self._lock:
-            if self._interrupted:
-                return False
-            else:
-                return self._event.is_set()
-
-    def clear_abort(self):
-        """Clears the abort.
-
-        """
-        with self._lock:
-            if self._interrupted:
-                self._interrupted = False
-                self._event.clear()
-
-
-class TimestampedData:
-    """ Class for storing timestamped data.
-
-    Attributes
-    ----------
-    timestamp : number-like
-        Timestamp associated with data.
-    data : object
-        Data to be stored.
-
-    """
-    def __init__(self, timestamp, data):
-        self.timestamp = timestamp
-        self.data = data
-
-    def update_from_csv(self, input_string):
-        """Update from comma-separated string, only if timestamp is newer.
-
-        Parameters
-        ----------
-        input_string : string
-            Comma-separated string. First value is timestamp, rest is data.
-
-        """
-        floats = string_to_floats(input_string)
-
-        if (len(floats) - 1) != len(self.data):
-            raise ValueError('Cannot update TimestampedData with incompatible data.')
-
-        if floats[0] > self.timestamp:
-            self.timestamp = floats[0]
-            self.data = floats[1:]
-
-    def update_from_data(self, timestamp, data):
-        """Update with data if timestamp is newer.
-
-        Parameters
-        ----------
-        timestamp : number-like
-            Timestamp associated with data.
-        data : object
-            Data to be stored if timestamp is newer.
-
-        """
-        if timestamp > self.timestamp:
-            self.timestamp = timestamp
-            self.data = data
-
-    @classmethod
-    def zeros(cls, length):
-        """ Construct empty TimestampedData object of specified length.
-
-        Parameters
-        ----------
-        length : int
-            Length of data to construct.
-
-        Return
-        ------
-        TimestampedData object
-
-        """
-        return cls(0, [0.] * length)
-
-    def __eq__(self, other):
-        """ Return true if other object has identical timestamp and data.
-
-        Parameters
-        ----------
-        other : object
-            Object to compare against.
-
-        Return
-        ------
-        bool
-            True if objects have same timestamp and data.
-
-        """
-        return other.timestamp == self.timestamp and other.data == self.data
-
-    def __ne__(self, other):
-        """ Return true if other object has different timestamp or data.
-
-        Parameters
-        ----------
-        other : object
-            Object to compare against.
-
-        Return
-        ------
-        bool
-            True if objects have different timestamp or data.
-
-        """
-        return not self == other
-
-
-class Message:
-    """Class for storing the internal state of a generic Mecademic robot.
-
-    Attributes
-    ----------
-    id : integer
-        The id of the message, representing the type of message.
-    data : string
-        The raw payoad of the message.
-
-    """
-    def __init__(self, id, data):
-        self.id = id
-        self.data = data
-
-    def __repr__(self):
-        return "Message with id={}, data={}".format(self.id, self.data)
 
 
 class RobotInfo:
@@ -688,175 +450,6 @@ class CallbackQueue():
         return self._queue.get(block=block, timeout=timeout)
 
 
-class CSVFileLogger:
-    """Class to handle logging robot state to file.
-
-    Attributes
-    ----------
-    file : file handle
-        File to be written to.
-    fields : list of strings
-        Fields to be logged.
-    command_queue : queue
-        Queue to store sent commands.
-    element_width : int
-        Each numerical element will have this width.
-
-    """
-    def __init__(self, robot_info, robot_state, fields=None, file_path=None, record_time=True):
-        """Initialize class.
-
-        Parameters
-        ----------
-        robot_info : RobotInfo
-            Contains robot information.
-        fields : list of strings
-            List of fields to be logged.
-        robot_state : RobotState
-            Contains state of robot.
-        file_path : string or None
-            If not provided, file will be saved in working directory.
-        record_time : bool
-            If true, current time will also be recorded in the text file. (Time is also available in filename.)
-
-        """
-        current_date_time = time.strftime('%Y-%m-%d-%H-%M-%S')
-
-        serial_number_or_blank = ('_serial_' + robot_info.serial) if robot_info.serial else ""
-
-        # Add unique name to file path.
-        file_name = (f"{robot_info.model}_R{robot_info.revision}_"
-                     f"v{robot_info.fw_major_rev}_{robot_info.fw_minor_rev}_{robot_info.fw_patch_num}_"
-                     f"log_{current_date_time}{serial_number_or_blank}.csv")
-
-        if file_path:
-            file_name = os.path.join(file_path, file_name)
-
-        # If fields argument is None, log all compatible fields.
-        if fields == None:
-            fields = []
-            for attr in vars(robot_state):
-                if attr.startswith('target') or attr.startswith('drive'):
-                    fields.append(attr)
-
-        # Set attributes.
-        self.file = open(file_name, 'w', newline='')
-        self.fields = fields
-        self.command_queue = queue.Queue()
-        self.element_width = 10
-
-        # Write robot information.
-        self.file.write('ROBOT_INFORMATION\n')
-        for attr in ['model', 'revision', 'fw_major_rev', 'fw_minor_rev', 'fw_patch_num']:
-            self.file.write(f'{attr}, {getattr(robot_info, attr)}\n')
-        if robot_info.serial != None:
-            self.file.write(f'serial_number, {robot_info.serial}\n')
-        if record_time:
-            self.file.write(f'time_recorded, {current_date_time}\n')
-
-        # Write headers for logged data.
-        self.file.write('\nLOGGED_DATA\n')
-        self.write_field_headers(robot_state)
-        self.write_field_and_element_headers(robot_info)
-
-    def write_field_headers(self, robot_state):
-        """For a field with multiple elements, write field name in first column, empty spaces for rest.
-
-        These headers are mostly for CSV and excel legibility.
-
-        Parameters
-        ----------
-        robot_state : RobotState
-            Current state of robot. Used only to get length of data fields.
-
-        """
-        self.file.write(f"{'timestamp':>15},")
-        for field in self.fields:
-            # Get number of elements in each field.
-            num_elements = len(getattr(robot_state, field).data)
-
-            # Add appropriate number of commas to align columns.
-            commas = ',' * (num_elements - 1)
-
-            # Calculate width of field given number of elements, accounting for commas.
-            width = (self.element_width + 1) * num_elements - 1
-            self.file.write(f'{field + commas:{width}},')
-        self.file.write('\n')
-
-    def write_field_and_element_headers(self, robot_info):
-        """Write the full field name and element name in each column.
-
-        Parameters
-        ----------
-        robot_info : RobotInfo
-            Information about the robot, such as model name and number of joints.
-
-        """
-        def assemble_with_prefix(field, names):
-            return ','.join([field + '_' + str(x) for x in names]) + ','
-
-        # Write full name for each field.
-        self.file.write(f"{'timestamp':>15},")
-        for field in self.fields:
-            if (field.endswith('joint_positions') or field.endswith('joint_velocity')
-                    or field.endswith('joint_torque_ratio')):
-                # Write field name followed by joint number. For example: "target_joint_positions_1".
-                self.file.write(assemble_with_prefix(field, range(robot_info.num_joints)))
-            elif field.endswith('end_effector_pose'):
-                self.file.write(assemble_with_prefix(field, ['x', 'y', 'z', 'alpha', 'beta', 'gamma']))
-            elif field.endswith('end_effector_velocity'):
-                self.file.write(
-                    assemble_with_prefix(field, ['x_dot', 'y_dot', 'z_dot', 'omega_x', 'omega_y', 'omega_z']))
-            elif field.endswith('configurations'):
-                self.file.write(assemble_with_prefix(field, ['shoulder', 'elbow', 'wrist']))
-            elif field.endswith('last_joint_turn'):
-                self.file.write(field + ',')
-            else:
-                raise ValueError(f'Missing formatting for field: {field}')
-        self.file.write('\n')
-
-    def write_fields(self, timestamp, robot_state):
-        """Write fields to file.
-
-        Parameters
-        ----------
-        timestamp : numeric
-            The timestamp of the current data.
-        robot_state : RobotState
-            This object contains the current robot state.
-
-        """
-        if self.file.closed:
-            return
-
-        # First write the timestamp
-        self.file.write(f'{timestamp:15},')
-
-        for field in self.fields:
-            # For each field, write each value with appropriate spacing.
-            self.file.write(','.join([f'{x:{self.element_width}}' for x in getattr(robot_state, field).data]))
-            self.file.write(',')
-
-        # End line with newline.
-        self.file.write('\n')
-
-    def end_log(self, ignore_checkpoints=True):
-        """Write all accumulated sent commands and close file.
-
-        """
-        # Write all sent commands.
-        self.file.write('\nSENT_COMMANDS\n')
-        while not self.command_queue.empty():
-            command = self.command_queue.get()
-
-            if ignore_checkpoints and command.startswith('SetCheckpoint'):
-                continue
-
-            self.file.write(f'"{command}"\n')
-
-        self.file.close()
-
-
 def disconnect_on_exception(func):
     """Decorator to call disconnect if an exception is raised. Needs to be declared outside of class.
 
@@ -879,23 +472,6 @@ def disconnect_on_exception(func):
                 raise e
 
     return wrap
-
-
-def string_to_floats(input_string):
-    """Convert comma-separated floats in string form to list of floats.
-
-    Parameters
-    ----------
-    input_string : string
-        Comma-separated floats values encoded as a string.
-
-    Returns
-    -------
-    list of floats
-        Returns converted list of floats.
-
-    """
-    return [float(x) for x in input_string.split(',')]
 
 
 class Robot:
